Move unit tests to pre-exist file
